Move test data to new folder
