+++ src/app/services/chat/chat.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from '@angular/core/testing';
 import { Subscription } from 'rxjs';
 import Chat from 'src/app/models/chat';
-import { SampleUsers } from '../user/sampleUsers';
+import { SampleUsers } from '../../test/data/sampleUsers';
 import { UserService } from '../user/user.service';
 
 import { ChatService } from './chat.service';
