+++ src/app/services/user/user.service.spec.ts
@@ -1,7 +1,7 @@
 import { fakeAsync, TestBed, tick } from '@angular/core/testing';
 import User from 'src/app/models/user';
 import { WishService } from '../wish/wish.service';
-import { SampleUsers } from './sampleUsers';
+import { SampleUsers } from '../../test/data/sampleUsers';
 import { UserService } from './user.service';
 
 describe('UserService', () => {
